:sparkles: Feat: add model classes for validating form and transferring data

+++ authentication/authentication-data/src/main/java/com/sowhat/authentication_data/model/response/NewMemberInfo.kt
@@ -9,5 +9,5 @@ data class NewMemberInfo(
     @SerialName("accessToken")
     val accessToken: String,
     @SerialName("memberId")
-    val memberId: Int
+    val memberId: Long
 )
